+++ src/components/LiquiditySelector/useCurrentPriceFromTicks.ts
@@ -6,7 +6,7 @@ import { tickIndexToPrice } from '../../lib/web3/utils/ticks';
 
 // current price of A to B is given in price B/A
 // eg. price of ATOM in USDC is given in USDC/ATOM units
-export default function useCurrentPriceFromTicks(
+export function useCurrentPriceFromTicks(
   tokenA?: string,
   tokenB?: string
 ): BigNumber | undefined {
@@ -18,17 +18,21 @@ export default function useCurrentPriceFromTicks(
   }, [midIndex]);
 }
 
-function useMidTickIndexFromTicks(
+export default function useMidTickIndexFromTicks(
   tokenA?: string,
   tokenB?: string
 ): number | undefined {
-  const { data: { ticks = [], token0, token1 } = {} } = useIndexerPairData(
-    tokenA,
-    tokenB
-  );
+  const {
+    data: {
+      token0Ticks: [token0Tick] = [],
+      token1Ticks: [token1Tick] = [],
+      token0,
+      token1,
+    } = {},
+  } = useIndexerPairData(tokenA, tokenB);
 
   // skip if there are no ticks to gain insight from
-  if (!ticks.length) {
+  if (!token0Tick && !token1Tick) {
     return;
   }
 
@@ -40,38 +44,12 @@ function useMidTickIndexFromTicks(
     return;
   }
 
-  const sortedTicks = ticks
-    // ignore irrelevant ticks
-    .filter((tick) => !tick.reserve0.isZero() || !tick.reserve1.isZero())
-    // ensure ticks are sorted
-    .sort((a, b) => a.tickIndex.comparedTo(b.tickIndex));
-
-  const highestTick0 = sortedTicks.reduceRight<TickInfo | undefined>(
-    (result, tick) => {
-      if (result === undefined && tick.reserve0.isGreaterThan(0)) {
-        return tick;
-      }
-      return result;
-    },
-    undefined
-  );
-
-  const lowestTick1 = sortedTicks.reduce<TickInfo | undefined>(
-    (result, tick) => {
-      if (result === undefined && tick.reserve1.isGreaterThan(0)) {
-        return tick;
-      }
-      return result;
-    },
-    undefined
-  );
-
   const midTickIndex =
-    highestTick0 && lowestTick1
+    token0Tick && token1Tick
       ? // calculate mid point
-        getMidIndex(highestTick0, lowestTick1)
+        getMidIndex(token0Tick, token1Tick)
       : // or return only found side of liquidity
-        highestTick0?.tickIndex ?? lowestTick1?.tickIndex;
+        token0Tick?.tickIndex ?? token1Tick?.tickIndex;
 
   return (reverse ? midTickIndex?.negated() : midTickIndex)?.toNumber();
 
@@ -84,19 +62,8 @@ function useMidTickIndexFromTicks(
       return highestTick0.tickIndex;
     }
     // linearly interpolate an answer
-    // get weights of each side in terms of token0 units
-    const highestTick0Value = sortedTicks
-      .filter((tick) => tick.tickIndex.isEqualTo(highestTick0.tickIndex))
-      .reduce((result, tick) => {
-        return result.plus(tick.reserve0);
-      }, new BigNumber(0))
-      .dividedBy(Math.pow(1.0001, highestTick0.tickIndex.toNumber()));
-    const lowestTick1Value = sortedTicks
-      .filter((tick) => tick.tickIndex.isEqualTo(lowestTick1.tickIndex))
-      .reduce((result, tick) => {
-        return result.plus(tick.reserve1);
-      }, new BigNumber(0))
-      .multipliedBy(Math.pow(1.0001, lowestTick1.tickIndex.toNumber()));
+    const highestTick0Value = highestTick0.reserve0;
+    const lowestTick1Value = lowestTick1.reserve1;
     // calculate the mid point
     const linearPercentage = lowestTick1Value.dividedBy(
       highestTick0Value.plus(lowestTick1Value)