+++ src/lib/web3/utils/ticks.ts
@@ -5,9 +5,14 @@ export function tickIndexToPrice(tickIndex: BigNumber): BigNumber {
   return new BigNumber(Math.pow(1.0001, tickIndex.toNumber()));
 }
 
-export function priceToTickIndex(price: BigNumber): BigNumber {
+export function priceToTickIndex(
+  price: BigNumber,
+  roundingMethod = 'none' as 'round' | 'ceil' | 'floor' | 'none'
+): BigNumber {
+  const roundingFunction =
+    roundingMethod !== 'none' ? Math[roundingMethod] : (v: number) => v;
   return new BigNumber(
-    Math.round(Math.log(price.toNumber()) / Math.log(1.0001))
+    roundingFunction(Math.log(price.toNumber()) / Math.log(1.0001))
   );
 }